frontend/requirements.txt
python-dotenv
streamlit
streamlit-lottie
boto3
streamlit_extras
plotly
matplotlib
google-cloud-storage
